refactor: migrate texture
